pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "fls-training"
version = "0.1.0"
description = "AI-powered FLS surgical training video scoring and feedback"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "openai>=1.50.0",
    "opencv-python-headless>=4.9.0",
    "ffmpeg-python>=0.2.0",
    "pydantic>=2.6.0",
    "duckdb>=1.1.0",
    "pyyaml>=6.0",
    "rich>=13.7.0",
    "python-dotenv>=1.0.0",
    "click>=8.1.0",
    "Pillow>=10.0.0",
]

[project.optional-dependencies]
api = ["fastapi>=0.110.0", "uvicorn>=0.27.0"]
training = [
    "torch>=2.2.0",
    "transformers>=4.45.0",
    "peft>=0.13.0",
    "accelerate>=0.34.0",
    "bitsandbytes>=0.44.0",
    "datasets>=3.0.0",
    "wandb>=0.18.0",
]
dev = ["pytest>=8.0.0", "ruff>=0.3.0"]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.ruff]
target-version = "py311"
line-length = 100
